+++ CHANGELOG.md
@@ -2,6 +2,52 @@
 
 All notable changes to this project will be documented in this file.
 
+### [5.1.2](https://github.com/wandb/terraform-azurerm-wandb/compare/v5.1.1...v5.1.2) (2025-01-10)
+
+
+### Bug Fixes
+
+* Helm cursed comma ([#117](https://github.com/wandb/terraform-azurerm-wandb/issues/117)) ([a405ed0](https://github.com/wandb/terraform-azurerm-wandb/commit/a405ed0af7458dced02c54f43d2855087fd596bd))
+
+### [5.1.1](https://github.com/wandb/terraform-azurerm-wandb/compare/v5.1.0...v5.1.1) (2025-01-09)
+
+
+### Bug Fixes
+
+* Identity typo ([#116](https://github.com/wandb/terraform-azurerm-wandb/issues/116)) ([5c46eea](https://github.com/wandb/terraform-azurerm-wandb/commit/5c46eeaf99920eb32e97b66d1c7c9e6f1326b33f))
+
+## [5.1.0](https://github.com/wandb/terraform-azurerm-wandb/compare/v5.0.0...v5.1.0) (2025-01-08)
+
+
+### Features
+
+* Pass console identity info ([#115](https://github.com/wandb/terraform-azurerm-wandb/issues/115)) ([d945e4c](https://github.com/wandb/terraform-azurerm-wandb/commit/d945e4cea3d623382283d23d24d820332b2ba18b))
+
+## [5.0.0](https://github.com/wandb/terraform-azurerm-wandb/compare/v4.1.2...v5.0.0) (2025-01-07)
+
+
+### ⚠ BREAKING CHANGES
+
+* Bump helm-wandb vers to v2.0.0, requires TF > 1.9 (#114)
+
+### Features
+
+* Bump helm-wandb vers to v2.0.0, requires TF > 1.9 ([#114](https://github.com/wandb/terraform-azurerm-wandb/issues/114)) ([77b7639](https://github.com/wandb/terraform-azurerm-wandb/commit/77b763934f52875af7fb77fbc102c49c21bc7d33))
+
+### [4.1.2](https://github.com/wandb/terraform-azurerm-wandb/compare/v4.1.1...v4.1.2) (2024-12-18)
+
+
+### Bug Fixes
+
+* Tags were missing on redis ([#113](https://github.com/wandb/terraform-azurerm-wandb/issues/113)) ([48afa7c](https://github.com/wandb/terraform-azurerm-wandb/commit/48afa7ca3fbbcd1a1308fe7dd3f4055bac8fd503))
+
+### [4.1.1](https://github.com/wandb/terraform-azurerm-wandb/compare/v4.1.0...v4.1.1) (2024-12-17)
+
+
+### Bug Fixes
+
+* Don't put special tags on the DB, added this to the var we pass in ([#112](https://github.com/wandb/terraform-azurerm-wandb/issues/112)) ([7f5c51d](https://github.com/wandb/terraform-azurerm-wandb/commit/7f5c51d6d48d173fd1c1542ac6e28d63395dea94))
+
 ## [4.1.0](https://github.com/wandb/terraform-azurerm-wandb/compare/v4.0.0...v4.1.0) (2024-12-04)